Add Relax NG schema and sample encodings.
The schema is based on the current `tei-lex-0` proposal. It adds the
value `collocation` to `cit/@type` to enable direct encoding
of collocations.

The sample articles are the ones discussed in  the session on lexical
modeling, i. e. _aid_ and _approach_ from the dictionary of basic
English we looked at.